+++ tests/Integration/ReplacerTest.php
@@ -30,7 +30,7 @@ class ReplacerTest extends TestCase
     }
 
     /**
-     * Test the error replacer for the 'postal_code_for' rule.
+     * Test the error replacer for the 'postal_code_with' rule.
      *
      * @return void
      */
@@ -41,11 +41,11 @@ class ReplacerTest extends TestCase
 
         $validator = $this->app->make('validator')->make(
             ['postal_code' => 'not-a-postal-code', 'country' => 'NL'],
-            ['postal_code' => 'postal_code_for:country']
+            ['postal_code' => 'postal_code_with:country']
         );
 
         $translator->addLines([
-            'validation.postal_code_for' => ':attribute invalid, should be a :countries postal code (e.g. :examples)',
+            'validation.postal_code_with' => ':attribute invalid, should be a :countries postal code (e.g. :examples)',
         ], $locale);
 
         $this->assertContains(