+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 4.0.0
 
 - Removed support for PHP 7.2
+- Removed the `postal_code_for` validation rule